backend/ai_engine/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attrs==24.3.0
audioread==3.0.1
black==24.10.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
faiss-cpu==1.9.0.post1
fastapi==0.100.0
ffmpeg-python==0.2.0
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.8
langchain-experimental==0.0.61
langchain-text-splitters==0.2.1
langsmith==0.1.81
llvmlite==0.43.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mccabe==0.7.0
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
orjson==3.10.14
packaging==24.2
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==1.10.21
pyflakes==3.2.0
pyparsing==3.2.1
pypdfium2==4.30.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytubefix==8.12.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
setuptools==75.8.0
sniffio==1.3.1
soundfile==0.13.0
SQLAlchemy==2.0.37
starlette==0.27.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.23.1
yarl==1.18.3
youtube-transcript-api==0.6.3
